pyproject.toml
[tool.poetry]
name = "omnilife-framework-automations"
version = "0.1.0"
description = ""
authors = ["Alexandre Osada <alexandre.osada.dev@gmail.com>"]
readme = "README.md"

packages = [
    { include = "omnilife_framework_automations/project", from = "./components" },
    { include = "omnilife_framework_automations/notion", from = "./components" },
    { include = "omnilife_framework_automations/logger", from = "./components" },
    { include = "omnilife_framework_automations/google", from = "./components" },
    { include = "omnilife_framework_automations/event", from = "./components" },
    { include = "omnilife_framework_automations/parameter", from = "./components" },
    { include = "omnilife_framework_automations/task", from = "./components" },
    { include = "omnilife_framework_automations/infrastructure", from = "./components" },
    { include = "omnilife_framework_automations/autofill_urgent_project_lambda", from = "./bases" },
    { include = "omnilife_framework_automations/plan_next_week_lambda", from = "./bases" },
]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.34.50"
requests = "^2.31.0"
pendulum = "^3.0.0"
injector = "^0.21.0"
google-auth = "^2.28.2"
google-api-python-client = "2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
python-dotenv = "^1.0.1"
pytest-cov = "^4.1.0"
requests-mock = "^1.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
